Here's everything for the data center page.
Layout is a little scuffed, but hold it up side by side with the wireframe, and it's pretty much identical. I don't know where to find the font they wanted for the header text, so i found a close match. But other than a couple images I need to add to the images folder, this is everything.